% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul2one.R
\name{mul2one}
\alias{mul2one}
\title{align}
\usage{
mul2one(fas, template, ...)
}
\arguments{
\item{fas}{path to fasta files}

\item{template}{path to template}

\item{...}{parameters for \link[Biostrings:pairwiseAlignment]{pairwiseAlignment}.}
}
\value{
Returns a list of \link[Biostrings:PairwiseAlignments-class]{PairwiseAlignments} object.
}
\description{
alignment multiple fasta files to one template
}
\examples{
path <- system.file("extdata", package="multi2one")
fas <- dir(path, "^Seq", full.names = TRUE)
template <- file.path(path, "p5E_MCS_UbiP.fa")
mul2one(fas, template)
}
