% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2txt.R
\name{write2txt}
\alias{write2txt}
\title{save mul2one result}
\usage{
write2txt(mo, filename = "align.txt", num = 50)
}
\arguments{
\item{mo}{output of mul2one}

\item{filename}{output file name}

\item{num}{number of base per line}
}
\description{
save mul2one result to a text file
}
\examples{
path <- system.file("extdata", package="multi2one")
fas <- dir(path, "^Seq", full.names = TRUE)
template <- file.path(path, "p5E_MCS_UbiP.fa")
write2txt(mul2one(fas, template))
}
